change some things to functions instead of macros
This seems to be the way nom 5 wants you to do, not macro everything up.
In many ways, I feel like this will be much easier for me to reason
about anyways, so this is good. I may switch these back to macros at
some point since they're so trivial and their non-macro-form is so
wordy, but they're fine for now :)